Copies code from original project
It takes the original `src/` and `tests/` folders and puts them
into Gradle's prefered folder hierarchy.  It removes the placeholders
created by Gradle.

It rewrites the tests to replace JUnit4 + JMock with Spock, which is
more expressive.

The original project assumed you had dependencies like Guice already
installed.  It now specifies it in `build.gradle` so Gradle can
locate it and download it automatically.